Update deletion spec to use elasticsearch GC
With apologies to Brian, something we missed during the initial
investigation for this is that Elasticsearch already natively supports
this behavior; when versioned documents are deleted ElasticSearch
preserves them for a configurable amount of time before actually
removing them so that it can detect version conflicts.

The old design is maintained at the bottom of this document, but I've
added a section detailing the native support, and it's that behavior
that the implementation will use.

Change-Id: Ic9bffc2c6f2b3800eb19a65bcac97a462dff70e0