Support microversion for nova servers plugin
Nova v2 API has been deprecated for a while and
the codes will be removed in Newton. The current
v2.1 API is designed with microversion mechanism,
when nova adds new fields to nova resources, new
API microversions will be added for related APIs.

As a search engine, Searchlight should be able to
work with microversion and make use of all newly
added data fields.

Change-Id: I31b7ad3376c39f7909e028e71bf83508e4ec2408
Blueprint: support-microversion-for-nova