Retire the searchlight-spec project
As announced in openstack-discuss ML[1], Searchlight project
is retiring in Wallaby cycle.

This commit retires searchlight-spec repository as per process deinfed in
project-guide[2]. Anyone would like to maintain it again, please
revert back this commit and propose the re-adding it
to governance.

The community wishes to express our thanks and appreciation to all of
those who have contributed to the searchlight-spec project over the years.

Depends-On: https://review.opendev.org/c/openstack/project-config/+/764519
Needed-By: https://review.opendev.org/c/openstack/governance/+/764530

[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-November/018637.html
[2] https://docs.openstack.org/project-team-guide/repository.html#retiring-a-repository

Change-Id: Ibbe1d916be8047f2fe48be7b6b79ff8f9196662b